backend/requirements.txt
fastapi>=0.103.1
uvicorn>=0.23.2
sqlmodel>=0.0.8
pydantic>=2.4.2
pydantic-settings>=2.0.3
python-jose>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.6
email-validator>=2.0.0
loguru>=0.7.2
python-dotenv>=1.0.0
bcrypt>=4.0.1
alembic>=1.12.0
